Made Comment Controller

+++ app/Http/Controllers/Blog/PostController.php
@@ -1,28 +1,27 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Blog;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
+use App\Models\Comment;
 use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 
-use function PHPUnit\Framework\fileExists;
-
 class PostController extends Controller
 {
-    
     public function index()
     {
         $posts = Post::all();
 
-        return view('index', compact('posts'));
+        return view('posts.index', compact('posts'));
     }
 
     public function create()
     {
-        return view('create');
+        return view('posts.create');
     }
 
     public function store(StorePostRequest $request)
@@ -41,29 +40,32 @@ class PostController extends Controller
         }
 
 
-        return redirect('/')->with('success','Post Created Successfully');
+        return redirect()->route('post.index')->with('success','Post Created Successfully');
     }
 
-    public function show($id)
+    public function show(Post $post)
     {
-        $post = Post::findOrFail($id);
-        return view('show', compact('post'));
+        // $post = Post::findOrFail($id);
+        $comment = Comment::where('post_id', $post->id)->get();
+
+        return view('posts.show', compact('post','comment'));
     }
 
-    public function edit($id)
+    public function edit(Post $post)
     {
-        $post = Post::findOrFail($id);
-        return view('edit', compact('post'));
+        // $post = Post::findOrFail($id);
+        return view('posts.edit', compact('post'));
     }
 
-    public function update(UpdatePostRequest $request, $id)
+    public function update(UpdatePostRequest $request, Post $post)
     {
-        $post = Post::findOrFail($id);
-        
+        // $post = Post::findOrFail($id);
         if ($request->hasFile('image')) {
-            $destination = 'images/posts/'.$post->image;
-            if (File::exists($destination)) {
-                unlink($destination);
+            if ($post->image) {
+                $destination = 'images/posts/'.$post->image;
+                if (File::exists($destination)) {
+                    unlink($destination);
+                }
             }
             $file = $request->file('image');
             $extension = $file->getClientOriginalExtension();
@@ -75,12 +77,12 @@ class PostController extends Controller
             $post->update($request->validated());
         }
 
-        return redirect('/')->with('success','Post Updated Successfully');
+        return redirect()->route('post.index')->with('success','Post Updated Successfully');
     }
 
-    public function destroy($id)
+    public function destroy(Post $post)
     {
-        $post = Post::findOrFail($id);
+        // $post = Post::findOrFail($id);
 
         if ($post->image) {
             $destination = 'images/posts/'.$post->image;
@@ -91,6 +93,6 @@ class PostController extends Controller
 
         $post->delete();
 
-        return redirect('/')->with('success','Post '.$post->title.' Deleted Successfully');
+        return redirect()->route('post.index')->with('success','Post '.$post->title.' Deleted Successfully');
     }
 }
